Add files via upload
Made an excerpt from each description (tei:row) to give only the parts, where there is text i.e. grammar mistakes. The html can be copy-pasted in word or google docs and so use the spell check.
Work in progress! Included are only descriptions of fragments with noted "Bearbeiter" as of 02. August.